Progressing on OCS DOS method
These were developed on my desktop, which is not stable these days.
This commit is to transfer the work to `srv`.
