+++ src/brd/predict.jl
@@ -1,10 +1,10 @@
 """
-    predict!(ID::AbstractVector{T}, ped::DataFrame, trts::Trait...) where T <: Integer
+    Predict!(ID::AbstractVector{T}, ped::DataFrame, trts::Trait...) where T <: Integer
 
 Give random EBV of trait `trts` about ID specified with `ID` in DataFrame `ped`.
 The rest of the values, and the sex not specific values are set as missing.
 """
-function predict!(ID::AbstractVector{T}, ped::DataFrame, trts::Trait...) where T <: Integer
+function Predict!(ID::AbstractVector{T}, ped::DataFrame, trts::Trait...) where T <: Integer
     1 ≤ minimum(ID) ≤ maximum(ID) ≤ nrow(ped) || throw(ArgumentError("ID out of range"))
     for trt in trts
         ebv = "ebv_" * trt.name
@@ -15,7 +15,7 @@ function predict!(ID::AbstractVector{T}, ped::DataFrame, trts::Trait...) where T
 end
 
 """
-    predict!(ID::AbstractVector{T}, ped::DataFrame,
+    Predict!(ID::AbstractVector{T}, ped::DataFrame,
         fixed::Vector{AbstractString}, giv::Matrix{Float64},
         trts::Trait...) where T <: Integer
 
@@ -23,7 +23,7 @@ Predict the trait(s) `trts` EBV of `ID` in DataFrame `ped` with the fixed
 effects `fixed`, which is/are also column(s) in `ped`, the inverse relationship
 matrix `giv`.
 """
-function predict!(ID::AbstractVector{T}, ped::DataFrame,
+function Predict!(ID::AbstractVector{T}, ped::DataFrame,
     fixed::Vector{S}, giv::Matrix{Float64},
     trts::Trait...) where {T <: Integer, S <: AbstractString}
     1 ≤ minimum(ID) ≤ maximum(ID) ≤ nrow(ped) || throw(ArgumentError("ID out of range"))
@@ -37,14 +37,14 @@ end
 
 #=
 """
-    predict!(ped::DataFrame, trts::Trait...; all = false)
+    Predict!(ped::DataFrame, trts::Trait...; all = false)
 
 Give random values to the :ebv of trait `trts` in DataFrame `ped`. If `all` is
 true, or the ebv column the trait doesn't exist in `ped`, then all ID will be 
 assigned a random value. Otherwise, only those in the last generation will be
 assigned random values.
 """
-function predict!(ped::DataFrame, trts::Trait...; all = false)
+function Predict!(ped::DataFrame, trts::Trait...; all = false)
     for trt in trts
         #ex = Regex("\\Q$(trt.name)\\E\$") # to find all about a trait
         tgt = all ? ped : groupby(ped, :grt)[end]
@@ -53,11 +53,11 @@ function predict!(ped::DataFrame, trts::Trait...; all = false)
     ped
 end
 
-function predict!(ped::DataFrame, cf, giv::Matrix{Float64}, trts::Trait...;
+function Predict!(ped::DataFrame, cf, giv::Matrix{Float64}, trts::Trait...;
     all = false, h2 = :known)
 end
 
-function predict!(ID::AbstractVector{Int}, ped::DataFrame, cf, giv::Matrix{Float64},
+function Predict!(ID::AbstractVector{Int}, ped::DataFrame, cf, giv::Matrix{Float64},
     trts::Trait...; all = false, h2 = :known)
 end
 =#
