+++ src/xps.jl
@@ -15,12 +15,14 @@ using Serialization
 using Term
 using xyBnG.XY
 import xyBnG.xyTypes: Cattle, Species, Trait, Plan, name
-import xyBnG.Breeding: phenotype!, predict!, Select, reproduce!, TM1997
+import xyBnG.Breeding: phenotype!, Predict!, Select, reproduce!, TM1997
 import xyBnG.Founder: sim_base, sample_founder, uniq
 import xyBnG.Conn.TS: sample2xy
 import xyBnG.RS: nrm, irm
 
 include("xps/a-tale-of-selction.jl")
+include("xps/one-trait.jl")
 include("xps/devel.jl")
+include("xps/dosblup.jl")
 
 end # module xps
