+++ src/xyBnG.jl
@@ -11,7 +11,9 @@ include("xps.jl")
 
 import xyBnG.xyTypes: Species, Cattle, Trait
 import xyBnG.Founder: sim_base
+import xyBnG.xps: devel
 
 export Species, Cattle, Trait, sim_base
+export devel
 
 end # module xyBnG
